Refactored reducers
