+++ packages/ra-tree-core/src/index.js
@@ -1,5 +1,5 @@
 export TreeController from './TreeController';
 export TreeContext from './TreeContext';
-export reducer from './reducers';
+export reducer from './reducer';
 export * from './actions';
 export * from './selectors';
