--- packages/ra-tree-core/src/reducers/index.js
@@ -1,6 +1,0 @@
-import { combineReducers } from 'redux';
-import expanded from './expanded';
-
-export default combineReducers({
-    expanded,
-});
