+++ packages/ra-tree-core/src/selectors.js
@@ -1,15 +1,7 @@
 export const getIsNodeExpanded = (state, nodeId) => {
-    const nodeStatus = state.expanded[nodeId] || {
+    const nodeStatus = state[nodeId] || {
         isExpanded: false,
         fromHover: false,
     };
     return nodeStatus.isExpanded;
-};
-
-export const getIsNodeExpandedFromHover = (state, nodeId) => {
-    const nodeStatus = state.expanded[nodeId] || {
-        isExpanded: false,
-        fromHover: false,
-    };
-    return nodeStatus.fromHover;
 };
